+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Timesheet : Assign Task to lines when generating from Realise Planning. 
 - Project : Added new O2M field 'announcementList' in bussiness-support module.
 - TeamTask : Addition of new boolean 'isOrderProposed' in business-support module and label on isOrderAccepted.
+- Timesheet : Assign toInvoice while generating timesheetLine from Realise Planning.
 
 ## Improvements
 - TimesheetLine: Adding M2O field TeamTask and integer field timeToCharge.